==> code/03_plots.Rmd <==
---
title: "02 - Exploratory Plots"
author: "Lina Walkowiak"
date: "`r Sys.Date()`"
output: html_document
---
# Exploration plots
## Load packages and data 
```{r}
library(sf)
library(tmap)
library(dplyr)
library(readr)
library(tidyr)
library(ggplot2)
library(scales)
library(RColorBrewer) 
library(ggplot2)
library(patchwork)


# Load preprocessed data
census_sections <- readRDS("../data/processed/census_sections.rds")
before <- readRDS("../data/processed/before_final.rds")
after <- readRDS("../data/processed/after_final.rds")
spillover_combined <- readRDS("../data/processed/both_years.rds")

listings2015 <- readRDS("../data/processed/listings2015_sf.rds")
listings2025 <- readRDS("../data/processed/listings_sf.rds")


```

## Merge data and geometry 
```{r}
before_sf <- census_sections %>%
  left_join(st_drop_geometry(before), by = "census_id")

after_sf <- census_sections %>%
  left_join(st_drop_geometry(after), by = "census_id")

combined_sf <- census_sections %>%
  left_join(st_drop_geometry(combined), by = c("census_id" = "census_id"))

```


### Plots for income 
```{r}
# Create year tag and merge into one dataset
before_sf$year <- 2015
after_sf$year <- 2024
combined_sf <- bind_rows(before_sf, after_sf)

# Compute change in wide format
change_df <- combined_sf %>%
  st_drop_geometry() %>%
  select(census_id, year, Import_Euros) %>%
  tidyr::pivot_wider(names_from = year, values_from = Import_Euros, names_prefix = "income_") %>%
  mutate(
    raw_change = income_2024 - income_2015,
    z_change = scale(raw_change)[,1]  # normalised with z-score
  )

# Merge back with geometry
change_sf <- after_sf %>%
  left_join(change_df, by = "census_id")

```

```{r}

ggplot(change_sf) +
  geom_sf(aes(fill = z_change), color = "white", size = 0.05) +
  scale_fill_distiller(
    palette = "YlGnBu", 
    direction = 1,
    name = "Income change (z-score)",
    guide = guide_colorbar(barwidth = 1.2, barheight = 10)
  ) +
  labs(
    title = "Change in disposable income",
  ) +
  theme_minimal(base_size = 13) +
  theme(
    plot.title = element_text(size = 18, face = "bold", hjust = 0.5),
    legend.position = c(0.98, 0.05),
    legend.justification = c(0.5, 0),
    panel.grid = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank()
  )
```


# Education 
```{r}

# create grouping
edu_change_df <- combined_sf %>%
  st_drop_geometry() %>%
  mutate(
    high_edu = pct_edu_4 + pct_edu_5 + pct_edu_6
  ) %>%
  select(census_id, year, high_edu) %>%
  tidyr::pivot_wider(names_from = year, values_from = high_edu, names_prefix = "edu_") %>%
  mutate(
    raw_change = edu_2024 - edu_2015,
    z_change = scale(raw_change)[,1]
  )

# Merge back to 2024 geometry
edu_change_sf <- after_sf %>%
  left_join(edu_change_df, by = "census_id")

# main plot 
ggplot(edu_change_sf) +
  geom_sf(aes(fill = z_change), color = "white", size = 0.05) +
  scale_fill_gradientn(
    colours = pal,
    values = scales::rescale(c(-3, -2, -1, 0, 1, 3)),  # adapt to your range
    name = "Percentage of highly \n educated residents (z-score)",
    guide = guide_colorbar(barwidth = 1.2, barheight = 10)
  ) +
  labs(
    title = "Change in the share of highly educated residents"
  ) +
  theme_minimal(base_size = 13) +
  theme(
    plot.title = element_text(size = 18, face = "bold", hjust = 0.5),
    legend.position = c(0.98, 0.05),
    legend.justification = c(0.5, 0),
    panel.grid = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank()
  )

```

# listings 
```{r}
library(ggplot2)
library(sf)
library(patchwork)  # For combining plots

# Plot 2015 listings
p1 <- ggplot() +
  geom_sf(data = census_sections, fill = "grey95", color = "white") +
  geom_sf(data = listings2015, color = "#1b9e77", size = 0.2, alpha = 0.6) +
  labs(title = "Airbnb Listings (2015)") +
  theme_minimal(base_size = 13) +
  theme(
    plot.title = element_text(size = 18, face = "bold", hjust = 0.5),
    legend.position = c(0.98, 0.05),
    legend.justification = c(0.5, 0),
    panel.grid = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank()
  )

# Plot 2025 listings
p2 <- ggplot() +
  geom_sf(data = census_sections, fill = "grey95", color = "white") +
  geom_sf(data = listings2025, color = "#d95f02", size = 0.2, alpha = 0.6) +
  labs(title = "Airbnb Listings (2025)") +
  theme_minimal(base_size = 13) +
  theme(
    plot.title = element_text(size = 18, face = "bold", hjust = 0.5),
    legend.position = c(0.98, 0.05),
    legend.justification = c(0.5, 0),
    panel.grid = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank()
  )

# Combine them side by side
p1 + p2

```

# Spillover 
```{r}
# combied data containing spillover
spillover_combined <- st_as_sf(readRDS("../data/processed/both_years.rds"))


# Just to be safe, we drop geometry from spillover and use census
spillover_no_geom <- spillover_combined %>%
  st_drop_geometry()

# Join based on shared column
spillover_joined <- census_sections %>%
  left_join(spillover_no_geom, by = "census_id")

spillover_joined <- census_sections %>%
  left_join(spillover_no_geom, by = "census_id") %>%
  mutate(kernel_spill_300 = replace_na(kernel_spill_300, 0))


```


```{r}

# Filter spillover per year
spillover_2015 <- spillover_joined %>% filter(year == 2015)
spillover_2025 <- spillover_joined %>% filter(year == 2024)

#Clean listings
listings2015_clean <- listings2015 %>%
  filter(!st_is_empty(geometry)) %>%
  st_cast("POINT", warn = FALSE)

listings2025_clean <- listings2025 %>%
  filter(!st_is_empty(geometry)) %>%
  st_cast("POINT", warn = FALSE)


# normalise 
spillover_2015 <- spillover_2015 %>%
  mutate(kernel_spill_300_z = scale(kernel_spill_300))

spillover_2025 <- spillover_2025 %>%
  mutate(kernel_spill_300_z = scale(kernel_spill_300))


```

```{r}


ggplot() +
  geom_sf(data = spillover_2025, aes(fill = kernel_spill_300_z), color = NA) +
  geom_sf(data = listings2025_clean, color = "#00003f", size = 0.2, alpha = 0.3) +
   scale_fill_gradientn(
    colours = c("#d73027", "#fcdc83", "#ffffbf",  "#54c4b4", "#565fa4"),  
    values = scales::rescale(c(-3, -1.5, 0, 1.5, 3)),
    name = "Spillover (Z-score)",
    guide = guide_colorbar(barwidth = 1.2, barheight = 10),
    limits = c(-2.5, 4.5)
  ) +
  labs(title = "2025: Spillover and Listings (Z-score)") +
  theme_minimal() +
  theme(
    plot.title = element_text(size = 14, face = "bold", hjust = 0.5),
    legend.position = "right",
    panel.grid = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank()
  )


# 2015
ggplot() +
  geom_sf(data = spillover_2015, aes(fill = kernel_spill_300_z), color = NA) +
  geom_sf(data = listings2015_clean, color = "#00003f", size = 0.2, alpha = 0.3) +
   scale_fill_gradientn(
    colours = c("#d73027", "#fcdc83", "#ffffbf",  "#54c4b4", "#565fa4"),  
    values = scales::rescale(c(-3, -1.5, 0, 1.5, 3)),
    name = "Spillover (Z-score)",
    guide = guide_colorbar(barwidth = 1.2, barheight = 10),
    limits = c(-2.5, 4.5)
  ) +
  labs(title = "2015: Spillover and Listings (Z-score)") +
  theme_minimal() +
  theme(
    plot.title = element_text(size = 14, face = "bold", hjust = 0.5),
    legend.position = "right",
    panel.grid = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank()
  )

```
```{r}

ggplot() +
  geom_sf(data = spillover_2025, aes(fill = kernel_spill_300), color = NA) +
  geom_sf(data = listings2025_clean, color = "#00003f", size = 0.2, alpha = 0.3) +
   scale_fill_gradientn(
    colours = c("#d73027", "#fcdc83", "#ffffbf",  "#54c4b4", "#565fa4"),  
    values = scales::rescale(c(-3, -1.5, 0, 1.5, 3)),
    name = " Total Spillover ",
    guide = guide_colorbar(barwidth = 1.2, barheight = 10),
    limits = c(0, 1000)
  ) +
  labs(title = "2025: Spillover and Listings") +
  theme_minimal() +
  theme(
    plot.title = element_text(size = 14, face = "bold", hjust = 0.5),
    legend.position = "right",
    panel.grid = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank()
  )


# 2015
ggplot() +
  geom_sf(data = spillover_2015, aes(fill = kernel_spill_300), color = NA) +
  geom_sf(data = listings2015_clean, color = "#00003f", size = 0.2, alpha = 0.3) +
   scale_fill_gradientn(
    colours = c("#d73027", "#fcdc83", "#ffffbf",  "#54c4b4", "#565fa4"),  
    values = scales::rescale(c(-3, -1.5, 0, 1.5, 3)),
    name = "Total Spillover",
    guide = guide_colorbar(barwidth = 1.2, barheight = 10),
    limits = c(0, 110)
  ) +
  labs(title = "2015: Spillover and Listings") +
  theme_minimal() +
  theme(
    plot.title = element_text(size = 14, face = "bold", hjust = 0.5),
    legend.position = "right",
    panel.grid = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank()
  )
```
